pyproject.toml
[tool.poetry]
name = "plant_pathology_2020_fgvc7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9"
numpy = "^1.21.4"
Pillow = "^8.4.0"
tqdm = "^4.62.3"
matplotlib = "^3.4.3"
torch = "^1.10.0"
torchvision = "^0.11.1"
pandas = "^1.3.4"
sklearn = "^0.0"
opencv-python = "^4.5.4"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^21.10b0"
flake8 = "^4.0.1"
mypy = "^0.910"
ipykernel = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
# common
python_version = 3.9
show_column_numbers  = true
show_error_context = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports_per_module = true
# warning
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true